Move the MLflow wiring into example_utils and simplify the flag handling
hf_ptq.py is down from ~90 lines of tracking glue to three calls:
add_mlflow_args(parser), resolve_mlflow_args(args, parser) and
`with mlflow_run(args):`. Everything that knows about the flags, the
params, the tags and the summary paths now sits beside the other hf_ptq
helpers in example_utils.

Simplify how the URI is settled. Provenance is now just "did the user type
the flag", read before the fallback, so the whole thing is one assignment
and one branch:

    args.mlflow_required = args.mlflow is not None
    args.mlflow = args.mlflow or os.environ.get("MLFLOW_TRACKING_URI") or None

Dropping nargs="?" is what makes that work. The bare --mlflow form meant
"use $MLFLOW_TRACKING_URI", which is now what happens with no flag at all,
so it was redundant -- and it was the only source of the empty-string case
that forced the three-way logic. A bare --mlflow now gets argparse's own
"expected one argument".

Move the MLflow section of the README to the end, per review: it is a lot
of detail to meet before the feature examples like AutoQuantize. Added to
the contents table so it stays findable.
